requirements.txt
streamlit
numpy
pandas
joblib
scikit-learn
youtube_transcript_api
duckduckgo-search
requests
python-dotenv
